braves_webapp/requirements.txt
Django==5.1.2
djangorestframework==3.15.2
psycopg2-binary==2.9.9
django-pandas
pandas